services/learning_pipeline/requirements.txt
# Echo Brain Learning Pipeline Requirements

# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
asyncpg>=0.29.0  # PostgreSQL async driver
psycopg2-binary>=2.9.7  # PostgreSQL sync driver (fallback)
qdrant-client>=1.7.0  # Vector database client
pydantic>=2.5.0  # Data validation
pydantic-settings>=2.1.0  # Configuration management

# NLP and ML
sentence-transformers>=2.2.2  # For embeddings
transformers>=4.35.0  # Hugging Face transformers
torch>=2.1.0  # PyTorch (CPU version)
numpy>=1.24.0  # Numerical computing
scikit-learn>=1.3.0  # ML utilities

# Text processing
nltk>=3.8.1  # Natural language processing
spacy>=3.7.0  # Advanced NLP
beautifulsoup4>=4.12.0  # HTML parsing
markdown>=3.5.0  # Markdown processing
python-magic>=0.4.27  # File type detection

# Async and networking
aiofiles>=23.2.0  # Async file operations
aiohttp>=3.9.0  # Async HTTP client
httpx>=0.25.0  # Modern HTTP client
websockets>=12.0  # WebSocket support

# Database and caching
redis>=5.0.0  # Redis client
alembic>=1.13.0  # Database migrations
sqlalchemy[asyncio]>=2.0.0  # ORM with async support

# Configuration and logging
pyyaml>=6.0.1  # YAML configuration
python-dotenv>=1.0.0  # Environment variables
structlog>=23.2.0  # Structured logging
colorlog>=6.8.0  # Colored logging

# Monitoring and metrics
prometheus-client>=0.19.0  # Prometheus metrics
opentelemetry-api>=1.21.0  # Observability
opentelemetry-sdk>=1.21.0

# Security
cryptography>=41.0.0  # Encryption utilities
bcrypt>=4.1.0  # Password hashing

# Testing (dev/test environments)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
httpx[test]>=0.25.0

# Code quality (dev environment)
black>=23.11.0  # Code formatting
ruff>=0.1.0  # Linting and formatting
mypy>=1.7.0  # Static type checking
pre-commit>=3.6.0  # Git hooks

# Utilities
click>=8.1.7  # CLI framework
rich>=13.7.0  # Rich text and formatting
tqdm>=4.66.0  # Progress bars
python-dateutil>=2.8.2  # Date utilities
pathlib-abc>=0.1.1  # Path utilities

# Development and debugging
ipython>=8.18.0  # Interactive Python
ipdb>=0.13.13  # Debugger
memory-profiler>=0.61.0  # Memory profiling